src/api/requirements.txt
## Python 3.13.6 compatibility floors
fastapi>=0.116.1
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
# Migrate to psycopg v3 (binary wheel provides 3.13 support). Keep psycopg2-binary installed manually only if needed.
psycopg[binary]>=3.1.19
alembic>=1.13.2
python-multipart>=0.0.9
pyjwt>=2.6.0
python-jose[cryptography]>=3.3.0
cryptography>=42.0.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.3
pydantic-settings>=2.3.0
httpx>=0.27.0
pytest>=8.2.0
psycopg2-binary==2.9.10
